Move timer callbacks outside timer context 🤦‍♂️
From the FreeRTOS documentation (https://www.freertos.org/RTOS-software-timer.html):

***Important information on writing timer callback functions***
"Timer callback functions execute in the context of the timer service task. It is therefore essential that timer callback functions never attempt to block. For example, a timer callback function must not call vTaskDelay(), vTaskDelayUntil(), or specify a non zero block time when accessing a queue or a semaphore."