Enable polling mode for PCA9535 IO Expander!
Normally, whenever an input to the IO expander changes, it toggles the interrupt line to notify the processor. We didn't connect the interrupt line to the dev mote, so we can't tell whenever USB is connected/disconnected. Added a way to enable periodic polls to the IO expander to notice pin changes after they happen.