Allows Load Cells That Decrease Readings With Tension Applied
When tension is applied to a load cell that decreases in readings, the
logic will always zero out the reading. To prevent this, when the
factor/divider of the load cell is negative, the system will assume
there is negative readings. The factor/divider is a configuration in the
system partition named loadCellCalibrationFactor.