Fixes Bug When Default Keys Are Deleted Causing The Bridge To Crash
Re-orders how things are initialized to prevent receiving serial
messages before it is too early to allow for it. This allows the baud
rate to be set and negotiated properly after a commit to the
configuration partition.
Previously the baud rate would get re-negotiated before setting the
default configuration values and then upon the next reboot after saving
those configuration values, the Bridge would be at the wrong baud rate
and improperly try to parse packets at a much higher baud rate, causing
too many events to occur and the assert on line 676 of ncp_uart.cpp to
occur.