Improve Command Sending To Aanderaa 5990 Sensor
This improves the handling of commands sent to the Aanderaa 5990 sensor
from the mote application.
Each command waits until an acknowledge is received by the 5990 before
moving on, improving the speed at which the device can be configured.
Get commands are able to assign values generic values based on their outputs through
the checkTypeAndAssign overloaded functions.